2.2 info

+++ src/pl/edu/wsiz/io12/part2/Employee.java
@@ -3,5 +3,5 @@ package pl.edu.wsiz.io12.part2;
 public class Employee {
     String firstname;
     String lastname;
-    byte age;
+    int age;
 }
